+++ cli/src/commands/get/datasets.rs
@@ -1,5 +1,10 @@
+use std::collections::HashMap;
+
 use anyhow::{Context, Result};
-use reinfer_client::{Client, DatasetIdentifier};
+use log::info;
+use reinfer_client::{
+    resources::dataset::StatisticsRequestParams, Client, CommentFilter, DatasetIdentifier,
+};
 use structopt::StructOpt;
 
 use crate::printer::Printer;
@@ -9,11 +14,18 @@ pub struct GetDatasetsArgs {
     #[structopt(name = "dataset")]
     /// If specified, only list this dataset (name or id)
     dataset: Option<DatasetIdentifier>,
+
+    #[structopt(long = "stats")]
+    /// Whether to include source statistics in response
+    include_stats: bool,
 }
 
 pub fn get(client: &Client, args: &GetDatasetsArgs, printer: &Printer) -> Result<()> {
-    let GetDatasetsArgs { dataset } = args;
-    let datasets = if let Some(dataset) = dataset {
+    let GetDatasetsArgs {
+        dataset,
+        include_stats,
+    } = args;
+    let mut datasets = if let Some(dataset) = dataset {
         vec![client
             .get_dataset(dataset.clone())
             .context("Operation to list datasets has failed.")?]
@@ -26,5 +38,36 @@ pub fn get(client: &Client, args: &GetDatasetsArgs, printer: &Printer) -> Result
         });
         datasets
     };
+
+    let mut dataset_stats: HashMap<_, _> = HashMap::new();
+    if *include_stats {
+        datasets.iter().try_for_each(|dataset| -> Result<()> {
+            info!("Getting statistics for dataset {}", dataset.full_name().0);
+            let stats = client
+                .get_dataset_statistics(
+                    &dataset.full_name(),
+                    &StatisticsRequestParams {
+                        comment_filter: CommentFilter {
+                            reviewed:Some(reinfer_client::resources::comment::ReviewedFilterEnum::OnlyReviewed),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    },
+                )
+                .context("Could not get statistics for dataset")?;
+
+            dataset_stats.insert(dataset.id.clone(), stats);
+            Ok(())
+        })?;
+
+        datasets.iter_mut().for_each(|dataset| {
+            dataset.num_reviewed = if let Some(stats) = dataset_stats.get(&dataset.id).cloned() {
+                Some(*stats.num_comments)
+            } else {
+                None
+            }
+        });
+    }
+
     printer.print_resources(&datasets)
 }